requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
Pillow==10.0.0
google-generativeai==0.1.0
python-dotenv==1.0.0
drf-spectacular==0.26.4
whitenoise==6.5.0
django-allauth==0.54.0
python-decouple==3.8